+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+* `flexberry-simpleolv`, `flexberry-objectlistview`, `object-list-view` components:
+    * Fix `showLoadingTbodyClass` event.
 
 ## [0.7.0-beta.15] - 2016-10-10
 ### Changed